refactor(model): rename BaseModel to Model and List to ModelCollection
